+++ Sekoia.io/sekoiaio/operation_center/__init__.py
@@ -147,13 +147,13 @@ CreateNewBlock = type(
     },
 )
 
-CreatesNewCase = type(
-    "CreatesNewCase",
+CreateCase = type(
+    "CreateCase",
     (GenericAPIAction,),
     {
         "verb": "post",
         "endpoint": base_url + "cases",
-        "query_parameters": ["tags", "subscribers"],
+        "query_parameters": [],
     },
 )
 
@@ -194,18 +194,18 @@ GetCase = type(
     (GenericAPIAction,),
     {
         "verb": "get",
-        "endpoint": base_url + "cases/{case_uuid}",
+        "endpoint": base_url + "cases/{uuid}",
         "query_parameters": ["community_uuid"],
     },
 )
 
-UpdatesCase = type(
-    "UpdatesCase",
+UpdateCase = type(
+    "UpdateCase",
     (GenericAPIAction,),
     {
         "verb": "patch",
-        "endpoint": base_url + "cases/{case_uuid}",
-        "query_parameters": ["tags", "subscribers"],
+        "endpoint": base_url + "cases/{uuid}",
+        "query_parameters": [],
     },
 )
 
@@ -240,12 +240,12 @@ ListsAlertsAssociatedToCaseUuid = type(
     },
 )
 
-CreateNewCommentOnCase = type(
-    "CreateNewCommentOnCase",
+PostCommentOnCase = type(
+    "PostCommentOnCase",
     (GenericAPIAction,),
     {
         "verb": "post",
-        "endpoint": base_url + "cases/{case_uuid}/comments",
+        "endpoint": base_url + "cases/{uuid}/comments",
         "query_parameters": [],
     },
 )
@@ -307,14 +307,15 @@ AddEventsToACase = type(
     },
 )
 
-assets_base_url = "api/v2/asset-management/"
+assets_v1_base_url = "api/v1/asset-management/"
+assets_v2_base_url = "api/v2/asset-management/"
 
 ListTypesForAssets = type(
     "ListTypesForAssets",
     (GenericAPIAction,),
     {
         "verb": "get",
-        "endpoint": assets_base_url + "asset-types",
+        "endpoint": assets_v1_base_url + "asset-types",
         "query_parameters": [
             "match[uuid]",
             "match[name]",
@@ -331,7 +332,7 @@ ReturnsTypeForAssets = type(
     (GenericAPIAction,),
     {
         "verb": "get",
-        "endpoint": assets_base_url + "asset-types/{uuid}",
+        "endpoint": assets_v1_base_url + "asset-types/{uuid}",
         "query_parameters": [],
     },
 )
@@ -341,21 +342,20 @@ ListAssets = type(
     (GenericAPIAction,),
     {
         "verb": "get",
-        "endpoint": assets_base_url + "assets",
+        "endpoint": assets_v2_base_url + "assets",
         "query_parameters": [
             "limit",
             "offset",
-            "extended",
-            "match[community_uuid]",
-            "match[uuid]",
-            "match[name]",
-            "match[category_name]",
-            "match[category_uuid]",
-            "match[type_name]",
-            "match[type_uuid]",
-            "range[criticality]",
-            "date[created_at]",
-            "date[updated_at]",
+            "search",
+            "uuids",
+            "community_uuids",
+            "type",
+            "category",
+            "source",
+            "reviewed",
+            "rule_uuid",
+            "criticality",
+            "rule_version",
             "sort",
             "direction",
         ],
@@ -367,7 +367,17 @@ CreatesNewAsset = type(
     (GenericAPIAction,),
     {
         "verb": "post",
-        "endpoint": assets_base_url + "assets",
+        "endpoint": assets_v1_base_url + "assets",
+        "query_parameters": [],
+    },
+)
+
+CreatesNewAssetV2 = type(
+    "CreatesNewAssetV2",
+    (GenericAPIAction,),
+    {
+        "verb": "post",
+        "endpoint": assets_v2_base_url + "assets",
         "query_parameters": [],
     },
 )
@@ -377,7 +387,17 @@ DeletesAsset = type(
     (GenericAPIAction,),
     {
         "verb": "delete",
-        "endpoint": assets_base_url + "assets/{uuid}",
+        "endpoint": assets_v1_base_url + "assets/{uuid}",
+        "query_parameters": [],
+    },
+)
+
+DeletesAssetV2 = type(
+    "DeletesAsset",
+    (GenericAPIAction,),
+    {
+        "verb": "delete",
+        "endpoint": assets_v2_base_url + "assets/{uuid}",
         "query_parameters": [],
     },
 )
@@ -387,7 +407,7 @@ ReturnsAsset = type(
     (GenericAPIAction,),
     {
         "verb": "get",
-        "endpoint": assets_base_url + "assets/{uuid}",
+        "endpoint": assets_v1_base_url + "assets/{uuid}",
         "query_parameters": [],
     },
 )
@@ -397,7 +417,7 @@ ListsAttributesOfAssets = type(
     (GenericAPIAction,),
     {
         "verb": "get",
-        "endpoint": assets_base_url + "assets/{uuid}/attr",
+        "endpoint": assets_v1_base_url + "assets/{uuid}/attr",
         "query_parameters": ["limit", "offset"],
     },
 )
@@ -407,7 +427,7 @@ AddsAttributeToAsset = type(
     (GenericAPIAction,),
     {
         "verb": "post",
-        "endpoint": assets_base_url + "assets/{uuid}/attr",
+        "endpoint": assets_v1_base_url + "assets/{uuid}/attr",
         "query_parameters": ["name", "value"],
     },
 )
@@ -417,7 +437,7 @@ DeletesAttributeFromAsset = type(
     (GenericAPIAction,),
     {
         "verb": "delete",
-        "endpoint": assets_base_url + "assets/{uuid}/attr/{attribute_uuid}",
+        "endpoint": assets_v1_base_url + "assets/{uuid}/attr/{attribute_uuid}",
         "query_parameters": [],
     },
 )
@@ -427,7 +447,7 @@ ReturnsAttributeOfAsset = type(
     (GenericAPIAction,),
     {
         "verb": "get",
-        "endpoint": assets_base_url + "assets/{uuid}/attr/{attribute_uuid}",
+        "endpoint": assets_v1_base_url + "assets/{uuid}/attr/{attribute_uuid}",
         "query_parameters": [],
     },
 )
@@ -437,7 +457,7 @@ ListsKeysOfAssets = type(
     (GenericAPIAction,),
     {
         "verb": "get",
-        "endpoint": assets_base_url + "assets/{uuid}/keys",
+        "endpoint": assets_v1_base_url + "assets/{uuid}/keys",
         "query_parameters": ["limit", "offset"],
     },
 )
@@ -447,7 +467,7 @@ AddsKeyToAsset = type(
     (GenericAPIAction,),
     {
         "verb": "post",
-        "endpoint": assets_base_url + "assets/{uuid}/keys",
+        "endpoint": assets_v1_base_url + "assets/{uuid}/keys",
         "query_parameters": ["name", "value"],
     },
 )
@@ -457,7 +477,7 @@ DeletesKeyFromAsset = type(
     (GenericAPIAction,),
     {
         "verb": "delete",
-        "endpoint": assets_base_url + "assets/{uuid}/keys/{key_uuid}",
+        "endpoint": assets_v1_base_url + "assets/{uuid}/keys/{key_uuid}",
         "query_parameters": [],
     },
 )
@@ -467,7 +487,7 @@ ReturnsKeyOfAsset = type(
     (GenericAPIAction,),
     {
         "verb": "get",
-        "endpoint": assets_base_url + "assets/{uuid}/keys/{key_uuid}",
+        "endpoint": assets_v1_base_url + "assets/{uuid}/keys/{key_uuid}",
         "query_parameters": [],
     },
 )
@@ -477,7 +497,7 @@ ListsOwnersOfAssets = type(
     (GenericAPIAction,),
     {
         "verb": "get",
-        "endpoint": assets_base_url + "assets/{uuid}/owners",
+        "endpoint": assets_v1_base_url + "assets/{uuid}/owners",
         "query_parameters": ["limit", "offset"],
     },
 )
@@ -487,7 +507,7 @@ AddsOwnersToAsset = type(
     (GenericAPIAction,),
     {
         "verb": "post",
-        "endpoint": assets_base_url + "assets/{uuid}/owners",
+        "endpoint": assets_v1_base_url + "assets/{uuid}/owners",
         "query_parameters": ["owners"],
     },
 )
@@ -497,7 +517,7 @@ DeletesOwnerFromAsset = type(
     (GenericAPIAction,),
     {
         "verb": "delete",
-        "endpoint": assets_base_url + "assets/{uuid}/owners/{owner_uuid}",
+        "endpoint": assets_v1_base_url + "assets/{uuid}/owners/{owner_uuid}",
         "query_parameters": [],
     },
 )
@@ -507,7 +527,7 @@ ListsNamesForKeysOfAsset = type(
     (GenericAPIAction,),
     {
         "verb": "get",
-        "endpoint": assets_base_url + "attribute-names",
+        "endpoint": assets_v1_base_url + "attribute-names",
         "query_parameters": ["category", "type", "limit", "offset"],
     },
 )
@@ -517,7 +537,7 @@ ReturnsAttributeName = type(
     (GenericAPIAction,),
     {
         "verb": "get",
-        "endpoint": assets_base_url + "attribute-names/{uuid}",
+        "endpoint": assets_v1_base_url + "attribute-names/{uuid}",
         "query_parameters": [],
     },
 )
@@ -527,7 +547,7 @@ ListsCategoriesForAssets = type(
     (GenericAPIAction,),
     {
         "verb": "get",
-        "endpoint": assets_base_url + "categories",
+        "endpoint": assets_v1_base_url + "categories",
         "query_parameters": ["limit", "offset"],
     },
 )
@@ -537,7 +557,7 @@ ReturnsCategoryForAssets = type(
     (GenericAPIAction,),
     {
         "verb": "get",
-        "endpoint": assets_base_url + "categories/{uuid}",
+        "endpoint": assets_v1_base_url + "categories/{uuid}",
         "query_parameters": [],
     },
 )
@@ -547,7 +567,7 @@ ReturnsCategoryTypeForAssets = type(
     (GenericAPIAction,),
     {
         "verb": "get",
-        "endpoint": assets_base_url + "categories/{uuid}/types",
+        "endpoint": assets_v1_base_url + "categories/{uuid}/types",
         "query_parameters": ["limit", "offset"],
     },
 )